Move all logic out of main thread, to avoid ANR
Top 3 ANRs happen when try to require mutex in `MegaApiImpl` C++ class, it may not be a deadlock, just block over 5s, so Android treat it as ANR, by moving the MegaApi call out of main thread, we won't have ANR even this kind of block happens.
This change is somehow experimental, but it won't be worse than existing code, we can track ANR reports on Google developer console to see if this works.