Use a single threaded scheduler to avoid wrong behavior
Previously all logic are executed on main thread, which is serialized, because DownloadService store some state as instance variable, and when downloading multiple files in a folder, the same instance of DownloadService is used, so if we use io scheduler, duplicated parameters are passed to MegaApi, which causes some files are not downloaded. Using a single threaded scheduler is the simplest way to fix this bug.