AND-12841 - Homepage: Fix Referral Program Banner Multiple Clicking Issue
Add View.debounceClick() method in ViewUtils that takes an action as a lambda parameter. This extension method will prevent onClick actions from occurring multiple times due to rapid clicking from the user. MINIMUM_DEBOUNCE_INTERVAL specifies the interval as to when the clicks will be registered by the system.

Utilize debounceClick() on the banner and dismiss image in BannerAdapter.