Increase minSdkVersion from 21 to 23
AND-13201: Drop Android 5.x support

MEGA projects follow a "1% device distribution thread", which means that support for a particular Android version will be dropped when there are less than 1% of overall users using the app on that Android version. As of October 11, 2021, Google Play Console has reported a 0.21% and 0.47% app version utilization for both Android 5.0 and 5.1, respectively. Hence, support for Android Lollipop will be discontinued.

Increase the minSdkVersion to 23. In effect, Android Marshmallow is now the lowest supported OS by MEGA.