Remove ListenScrollChangesHelper and OnScrollChangeListenerAdapter
AND-13545: Maintenance - Remove ListenScrollChangesHelper

Remove `ListenScrollChangesHelper` from the codebase. The class was created to support versions lower than Android Marshmallow and since the minimum app OS is now Android Marshmallow, there is no need to keep the class around.

This also converts classes that use `addViewToListen()` into `setOnScrollChangeListener()` for `ScrollView` and `addOnScrollListener()` for `RecyclerView`, respectively.

Furthermore, remove `OnScrollChangeListenerAdapter` since it is related to `ListenScrollChangesHelper` and the class is no longer used by the app.