Use ContextCompat.startForegroundService() when starting UploadService
AND-13496: Crash: RemoteServiceException in ActivityThread.handleMessage (ActivityThread.java)

The exception occurs on devices running Android API Level 26 (Android Oreo) and above. Basically, the app calls the Service and waited 5 seconds for the Service to call startForeground(). In some instances, startForeground() is not called which results in an ANR and RemoteServiceException being thrown.

Beginning on Android API Level 26 (Android Oreo), the system imposes restrictions on creating services. When the app falls on that particular OS version, use Context.startForegroundService() to create a background Service and promote the Service to the foreground.

In order to handle compatibility with Android API Levels 23-25 (Android Marshmallow until Nougat), use ContextCompat.startForegroundService(). This will automatically configure the app to use Context.startService() when the app falls under those versions, and use Context.startForegroundService() when the app is under Android Oreo and above.

Other changes:
- Removed if (managerActivity != null) checking when creating an Intent to UploadService in ManagerActivity.
- Added null checking of getActivity() and info in QRCodeSaveBottomSheetDialogFragment.