Apply Backup Read-Only Restrictions to Kotlin-Converted LinksFragment and OutgoingSharesFragment
AND-13482 - Backup in File Manager Version 2 for Android

LinksFragment and OutgoingSharesFragment were converted to Kotlin due to changes in subtask AND-13751 (Remove call to onNodesSharedUpdate() when updateNodes()). The changes were basically migrated from an earlier commit where both classes were in Java.
When the user selects one or more Folders, The following options in the top Context Menu are hidden:

- Rename
- Move
- Move to Rubbish Bin
- Favourite
- Label
- Versions