AND-14645: Create Use Case to Move Business Account Logic out of SettingsCameraUploadsFragment
Currently, SettingsCameraUploadsFragment calls the method checkIfShouldShowBusinessCUAlert(), which directly calls megaApi.isBusinessAccount() and megaApi.isMasterBusinessAccount() to if the user's Business Account is allowed to enable Camera Uploads or not.

In order to conform to Clean Architecture, create Use Cases IsBusinessAccountActive and CheckEnableCameraUploadsStatus. IsBusinessAccountActive checks whether the user has an Active Business Account or not, while CheckEnableCameraUploadsStatus relies on GetAccountDetails and IsBusinessAccountActive to return the correct EnableCameraUploadsStatus, allowing callers to implement appropriate behavior depending on the status returned.

Additional Changes:
- Define isBusinessAccountActive() in MegaApiFacade, MegaApiGateway, AccountRepository and DefaultAccountRepository.
- Add ViewModelComponent::class compatibility in CameraUploadUseCases and TestCameraUploadUseCases.
- Modify TestAccountModule and TestCameraUploadUseCases in response to new Use Cases.
- Add Unit Tests for IsBusinessAccountActive in DefaultAccountRepositoryTest.
- Create DefaultCheckEnableCameraUploadsStatusTest to verify the Use Case behaviors.