AND-15169: Fix Sort By Icon Blinking Issue
The issue is caused because when the corresponding Fragment gets initialized again, SortByHeaderViewModel gets created again, resulting in timing issues. This causes the Sort By icon to blink in order to use the correct icon.

Instead of manually refreshing the Sort By header through adapter.notifyItemChanged(), SortByHeaderViewModel should be delegated with activityViewModels, so that the ViewModel will persist until the Activity gets destroyed.