AND-15169: Fix Content Not Displaying in Documents, Audio and Video Sections
Upon investigation, it was found that the content "disappeared" because whenever an update is received from the UI State of SortByHeaderViewModel, viewModel.refreshUi() is called which invalidated all the content, even though the user did not change the View Type from the start.

Some code from the initial MR were restored in order to quickly resolve the issue. Essentially, viewModel.refreshUi() will only be called on a explicit change to the View Type.