CC-3878: Fix Camera Uploads Sync Handles Synchronization on Successful Login
It is possible that the Login process will be finished, even though the Camera Uploads Sync Handles were not established. In order to ensure that the Sync Handles are set up before continuing the Login process, map the latest value of fetchNodesFlow() first before calling trySend() to notify external observers.