AND-15835: Fix Incoming and Outgoing Shares UI When Sorting in Grid View Mode
When the user is in IncomingSharesFragment or OutgoingSharesFragment and selects a different Sort Order in Grid View mode, the UI temporarily switches to List View before going back to Grid View with the new Sort Order applied.

This is due to the ManagerActivity functions refreshCloudOrder() and refreshOthersOrder(), which calls refreshSharesPageAdapter(). Calling this function creates new instances for IncomingSharesFragment, OutgoingSharesFragment and LinksFragment. As a result, both IncomingSharesFragment and OutgoingSharesFragment will begin with List View mode before switching to Grid View mode.

Changelog:
- Remove ManagerActivity calls to refreshCloudOrder() and refreshOthersOrder() in SortByBottomSheetDialogFragment.
- Configure affected Fragments to observe the LiveData SortByHeaderViewModel.orderChangeEvent, so that they can refresh their own content whenever a Sort Order change happens. This also prevents having to create new instances for IncomingSharesFragment, OutgoingSharesFragment and LinksFragment.