BAC-382: Fix Crash When Rotating the App In Backups
During screen rotation, ManagerActivity calls checkScrollElevation(), which subsequently calls InboxFragment.checkScroll(). The cause of the crash is that the binding was not initialized when checkScroll() is called.

Fix the crash by removing the double-bang operator (!!) in binding and using Null Safety to perform UI operations when binding is not null.