BAC-293: Fix Backed Up Files Not Being Displayed
When the User logs in to an account without Backups, logs out, logs in to an account with Backups, and goes to the Backups page, the content is not displayed.

This is because of MegaNodeUtilFacade.observeBackupFolder(), which has MonitorBackupFolder. When ManagerActivity is started, the Backups Folder is immediately observed and updated through that function. As MegaNodeUtilFacade is a Singleton, it will survive even though ManagerActivity is destroyed when the User logs out. When the User logs in, MonitorBackupFolder does not get called, resulting in the Backups Folder not being saved.

The issue is resolved by removing MegaNodeUtilFacade.observeBackupFolder() and injecting MonitorBackupFolder in ManagerViewModel. This ensures that MonitorBackupFolder begins observing and updating the Backups Folder whenever ManagerActivity is created, and its process destroyed when ManagerActivity is destroyed.

Changelog:
- Remove usage of MegaNodeUtilWrapper.observeBackupFolder() in ManagerActivity.
- Remove MegaNodeUtilWrapper.observeBackupFolder() and its implementation in MegaNodeUtilFacade.
- Remove all injected classes in MegaNodeUtilFacade.
- Inject MonitorBackupFolder in ManagerViewModel and set the received Backups Folder handle to MegaNodeUtil.myBackupHandle when receiving an update.