BAC-468: Fix Up to Date Heartbeats Not Being Sent when Camera Uploads Has Nothing to Upload
The issue is caused when the user opens the app, enables Camera Uploads and there is nothing to upload.

This is due to if (lastNodeHandle != cameraUploadRepository) in ReportUploadStatusUseCase. Since there is nothing to upload, the lastNodeHandle in the State remains -1, which prevented the Up to Date Heartbeats from being sent.

Changelog:
- Remove the "if" statement in ReportUploadStatusUseCase.
- Update ReportUploadStatusUseCaseTest.