BAC-695: Device Center - Simplify Backups Flow
Previously, accessing the Backups feature is done by selecting the "Backups" Dashboard Side Menu entry point.

In this rework, when the "DeviceCenter" Feature Flag is enabled, the "Backups" Dashboard Side Men entry point is removed. Accessing the Backups feature can only be done by going to Device Center, selecting a Backup Device and selecting a Backup Folder.

In order to accomplish the rework, some refactoring was made in BackupsFragment to make it compliant with Clean Architecture. The Back Stack Navigation logic in ManagerActivity was also revised in order to navigate back from Backups to Device Center.