AND-17888: Fix Crash When Opening a Node's Bottom Sheet Options
The cause of the issue was that the Node Thumbnail ImageView's parent is a ConstraintLayout. As such, calling the legacy setNodeThumbnail() caused the app to crash with a ClassCastException, because the function is expecting a RelativeLayout.

Instead of using the legacy setNodeThumbnail() in NodeOptionsBottomSheetDialogFragment, a ThumbnailRequest is used to load the Node Thumbnail.