BAC-797: Device Center - Fix Some Info Screen Options Not Working
Backup Folders are subject to read-only restrictions. This means that the "Move", "Rename" and "Move to the Rubbish Bin" Actions should be hidden in the File Info screen.

To fix the issue, use IsNodeInBackupsUseCase in GetAvailableNodeActionsUseCase, and hide the NodeActions MoveToRubbishBin, Rename and Move if the TypedNode passed is a Backup Node. Also, update the Tests in GetAvailableNodeActionsUseCaseTest and improve Test names for better readability.