AND-15063: Fix ForegroundServiceStartNotAllowedException on UploadUseCase
Beginning in Android 12, starting Foreground Services requires that the app is currently active, meaning that the app is in the Foreground and currently being used. The crash occurred because the Service that manually uploads a Folder was attempted to be started in the Background, which Android no longer allows.

In order to resolve the issue, modify UploadUseCase.upload() to add a check that the app is currently active before calling startForegroundService().