BAC-816: Cloud Drive - Improve How Media Discovery Is Being Accessed
The existing logic of showing and hiding the Media Discovery is convoluted and difficult to understand. These changes will make it easier to understand and maintain the Media Discovery visibility. This will also help with unifying the Back Stack behavior of Cloud Drive.

In addition, update the Test Names in FileBrowserViewModelTest to match the given convention.