BAC-817: Cloud Drive - Fix Media Discovery "Double Back" Issue
It was observed that when performing a Back Navigation from Media Discovery, the Media Discovery is dismissed but the current Nodes remained the same for a short while before being updated. This seemed like the User did a Back Navigation twice, which is not a great User Experience.

In order to fix the issue, the actions of updating of Cloud Drive nodes, removing the Media Discovery and going back to the previous view (in this case, Cloud Drive) must be synchronized using lifecycleScope.launch{}.