BAC-779: Fix Cloud Drive Not in Root Level When Clicking the Feature Again
This brings back the existing behavior wherein if the User is in Cloud Drive and clicks the Cloud Drive Navigation Button again, then the Nodes are shown from the Root Level. Media Discovery is also deactivated when this happens.

Furthermore, replace all null checks of MediaDiscoveryFragment with FileBrowserViewModel.isMediaDiscoveryOpen() to better determine if the User is in Media Discovery or not.