CU-286: Settings Camera Uploads - Move the New Video Compression Size Dialog to the ViewModel and UI State
All Business and Display Logic pertaining to the "Notify me when size is larger than" Dialog are moved to SettingsCameraUploadsViewModel and SettingsCameraUploadsState. This also fixes an issue wherein SettingsCameraUploadsFragment keeps getting initialized when changing the Screen Orientation.

Additional Tests are added in SettingsCameraUploadsViewModelTest to verify that the existing New Video Compression Size Dialog functionalities work as expected.