BAC-839: Fix IllegalStateException Crash in ManagerActivity.showMediaDiscovery()
The cause of the IllegalStateException is due to an attempt to commit a FragmentTransaction after the Activity's state has been saved in onSaveInstanceState(). When this happens, the FragmentTransaction is not remembered because it was never part of the Activity's state.

To fix the issue, ensure that ManagerActivity is currently active in removeFragment() before executing the FragmentTransaction.