CU-816: Fix Camera Uploads Uploading Duplicate Content
The issue stems in how the existence of the CameraUploadsRecord is being checked in GetNodeFromCloudUseCase. In the Use Case, it first calls GetNodeByOriginalFingerprintUseCase, which calls MegaApiGateway.getNodesByOriginalFingerprint(). The code retrieves the Node using the first element of the list, which is problematic because it is not guaranteed that the first element retrieved originated from the Camera Uploads Primary / Secondary Target Node. If the first element retrieved does not exist in the Target Node, Camera Uploads copies the Node to the Target Node, causing a duplicate upload.

To fix this, reorder the check in GetNodeFromCloudUseCase; first check if the CameraUploadsRecord exists in the Target Node. If it does not exist in the Target Node, check against all Folders. This eliminates any chance of duplicate uploads.

Detailed Changelog:

- Rename getNodeByOriginalFingerprint() into getNodesByOriginalFingerprint() in NodeRepository and change its return type to List<UnTypedNode>. Remove the retrieval of the first element to make the function generic.
- Rename getNodesFromFingerPrint() into getNodesByFingerprint() in NodeRepository.
- Rename GetNodeFromCloudUseCase to GetNodeFromCloudDriveUseCase.
- Create GetNodesByOriginalFingerprintUseCase to retrieve the list of all Nodes having the same Original Fingerprint by passing a null parentNodeId.
- Create GetNodesByFingerprintUseCase to retrieve the list of all Nodes having the same Fingerprint.
- Replace the first GetNodeByOriginalFingerprintUseCase with GetNodesByOriginalFingerprintUseCase in GetNodeFromCloudDriveUseCase.
- Replace all GetNodeByFingerprintAndParentNodeUseCase calls with GetNodesByFingerprintUseCase for the Fingerprint and Generated Fingerprint in GetNodeFromCloudDriveUseCase.
- Remove AddNodeType casting in GetNodesByFingerprintUseCase to prevent any situation that causes an early null return and causing Camera Uploads to upload the same File in the Target Folder/s.
- Remove GetNodeByOriginalFingerprintUseCase.
- Add and update Unit Tests.