CC-8164: Complete the ML Document Kit Scanning Process from Chat Screen
The new Document Scanner can be accessed by clicking the "Scan" Button in the Chat screen. Error Handling is included in all three entrypoints (Cloud Drive, File Contact List, Chat) by showing an Error Dialog when there is an error opening the ML Document Kit Scanner.

Add missing Tests to validate the Document Scanning error behavior in ManagerViewModelTest and ContactFileListViewModelTest. Tests are also added in ChatToolbarBottomSheetTest to validate that Analytics was sent when the respective Button is clicked.