CC-8556: Fix Empty Filename Allowing to Proceed when a Different File Type is Selected
The onTextChange of GenericTextField is triggered upon tapping the input text. The retrieved text contains the entire filename with the selected file extension. This happens before any file input manipulations are made.

Remove the selected file extension from the input before calling onFilenameChanged. If the file extension is the only one left in the filename, this ensures that the UI State retrieves an empty filename, and halting the document scanning process when clicking "Next".