CC-8532: Display the Entire Filename and Do Not Highlight the File Extension in the Scan Confirmation Page
Show the entire filename when the text input is in focus.

When in focus:
- the filename without the file extension is highlighted when it has the expected file extension.
- the entire filename is highlighted when it does not have the expected file extension.

Furthermore, display an error message if the filename does not end with the expected file extension (e.g. PDF is selected, but the filename is only "Scanned").

Lastly, include Additional Error Messages for "Missing File Extension Scenario".